+++ EntityLayer/Concrete/People/PersonLeave.cs
@@ -15,7 +15,7 @@ namespace EntityLayer.Concrete.People
         public SysLeaveType SysLeaveType { get; set; }
         public DateTime StartAt { get; set; }
         public DateTime EndAt { get; set; }
-        public short Status { get; set; }
+        public int Status { get; set; }
         public string LFile { get; set; }
         public string Note { get; set; }
         public int SyncID { get; set; }
